Cambios jpaController
ya cambie los jpa, por ejemplo eliminé CajaJpaController y
CajaJpaController1 hice un refactor name para que quede
CajaJpaController. y asi sucesivamente.
