--- src/Persistencia/BultoJpaController1.java
@@ -1,144 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Persistencia;
-
-import Persistencia.exceptions.NonexistentEntityException;
-import Persistencia.exceptions.PreexistingEntityException;
-import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
-import model.Bulto;
-
-/**
- *
- * @author daniel
- */
-public class BultoJpaController1 implements Serializable {
-
-    public BultoJpaController1(EntityManagerFactory emf) {
-        this.emf = emf;
-    }
-    private EntityManagerFactory emf = null;
-
-    public EntityManager getEntityManager() {
-        return emf.createEntityManager();
-    }
-
-    public void create(Bulto bulto) throws PreexistingEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            em.persist(bulto);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findBulto(bulto.getNroBulto()) != null) {
-                throw new PreexistingEntityException("Bulto " + bulto + " already exists.", ex);
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void edit(Bulto bulto) throws NonexistentEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            bulto = em.merge(bulto);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            String msg = ex.getLocalizedMessage();
-            if (msg == null || msg.length() == 0) {
-                int id = bulto.getNroBulto();
-                if (findBulto(id) == null) {
-                    throw new NonexistentEntityException("The bulto with id " + id + " no longer exists.");
-                }
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void destroy(int id) throws NonexistentEntityException {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            Bulto bulto;
-            try {
-                bulto = em.getReference(Bulto.class, id);
-                bulto.getNroBulto();
-            } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The bulto with id " + id + " no longer exists.", enfe);
-            }
-            em.remove(bulto);
-            em.getTransaction().commit();
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public List<Bulto> findBultoEntities() {
-        return findBultoEntities(true, -1, -1);
-    }
-
-    public List<Bulto> findBultoEntities(int maxResults, int firstResult) {
-        return findBultoEntities(false, maxResults, firstResult);
-    }
-
-    private List<Bulto> findBultoEntities(boolean all, int maxResults, int firstResult) {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            cq.select(cq.from(Bulto.class));
-            Query q = em.createQuery(cq);
-            if (!all) {
-                q.setMaxResults(maxResults);
-                q.setFirstResult(firstResult);
-            }
-            return q.getResultList();
-        } finally {
-            em.close();
-        }
-    }
-
-    public Bulto findBulto(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Bulto.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
-    public int getBultoCount() {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            Root<Bulto> rt = cq.from(Bulto.class);
-            cq.select(em.getCriteriaBuilder().count(rt));
-            Query q = em.createQuery(cq);
-            return ((Long) q.getSingleResult()).intValue();
-        } finally {
-            em.close();
-        }
-    }
-    
-}
