+++ src/Persistencia/CajeroJpaController.java
@@ -11,32 +11,32 @@ import java.io.Serializable;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
 import javax.persistence.Query;
-import model.Articulo;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
 import model.Cajero;
 
 /**
  *
  * @author daniel
  */
-public class CajeroJpaController implements Serializable{
-   
-   //constructor
+public class CajeroJpaController implements Serializable {
+    private static final long serialVersionUID = 1L;
+
     public CajeroJpaController(EntityManagerFactory emf) {
-    this.emf = emf;
-}
-    private EntityManagerFactory emf = null;
-    public CajeroJpaController() {
-   emf= Persistence.createEntityManagerFactory("ElMercaditoPU");
+        this.emf = emf;
     }
-    
+    private EntityManagerFactory emf = null;
+
+    CajeroJpaController() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
-//metodo de jpaController. Create
     public void create(Cajero cajero) throws PreexistingEntityException, Exception {
         EntityManager em = null;
         try {
@@ -45,7 +45,7 @@ public class CajeroJpaController implements Serializable{
             em.persist(cajero);
             em.getTransaction().commit();
         } catch (Exception ex) {
-            if (findCajero(cajero.getDni()) != null) {
+            if (findCajero(cajero.getNroCajero()) != null) {
                 throw new PreexistingEntityException("Cajero " + cajero + " already exists.", ex);
             }
             throw ex;
@@ -53,23 +53,22 @@ public class CajeroJpaController implements Serializable{
             if (em != null) {
                 em.close();
             }
-        
         }
     }
-//    metodo de jpaController. edit
+
     public void edit(Cajero cajero) throws NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            cajero= em.merge(cajero);
+            cajero = em.merge(cajero);
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                String id = cajero.getDni();
+                int id = cajero.getNroCajero();
                 if (findCajero(id) == null) {
-                    throw new NonexistentEntityException("The cajero with id " + id+ " no longer exists.");
+                    throw new NonexistentEntityException("The cajero with id " + id + " no longer exists.");
                 }
             }
             throw ex;
@@ -79,16 +78,6 @@ public class CajeroJpaController implements Serializable{
             }
         }
     }
-//    método de jpaController: obtener un elemento
-    public Cajero findCajero(String id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Cajero.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
 
     public void destroy(int id) throws NonexistentEntityException {
         EntityManager em = null;
@@ -98,9 +87,9 @@ public class CajeroJpaController implements Serializable{
             Cajero cajero;
             try {
                 cajero = em.getReference(Cajero.class, id);
-                cajero.getDni();
+                cajero.getNroCajero();
             } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The cajero with dni " + id + " no longer exists.", enfe);
+                throw new NonexistentEntityException("The cajero with id " + id + " no longer exists.", enfe);
             }
             em.remove(cajero);
             em.getTransaction().commit();
@@ -111,7 +100,6 @@ public class CajeroJpaController implements Serializable{
         }
     }
 
-    //método de jpa controller: obtener varios elementos
     public List<Cajero> findCajeroEntities() {
         return findCajeroEntities(true, -1, -1);
     }
@@ -123,7 +111,9 @@ public class CajeroJpaController implements Serializable{
     private List<Cajero> findCajeroEntities(boolean all, int maxResults, int firstResult) {
         EntityManager em = getEntityManager();
         try {
-            Query q = em.createQuery("select object(o) from Cajero as o");
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            cq.select(cq.from(Cajero.class));
+            Query q = em.createQuery(cq);
             if (!all) {
                 q.setMaxResults(maxResults);
                 q.setFirstResult(firstResult);
@@ -134,15 +124,30 @@ public class CajeroJpaController implements Serializable{
         }
     }
 
+    public Cajero findCajero(int id) {
+        EntityManager em = getEntityManager();
+        try {
+            return em.find(Cajero.class, id);
+        } finally {
+            em.close();
+        }
+    }
 
-//metodo de jpaController: obtener cantidad de elementos persistidos
     public int getCajeroCount() {
         EntityManager em = getEntityManager();
         try {
-            Query q = em.createQuery("select count(o) from Cajero as o");
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            Root<Cajero> rt = cq.from(Cajero.class);
+            cq.select(em.getCriteriaBuilder().count(rt));
+            Query q = em.createQuery(cq);
             return ((Long) q.getSingleResult()).intValue();
         } finally {
             em.close();
         }
-}
+    }
+
+    Cajero findCajero(String id) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+    
 }
