+++ src/Persistencia/CompraJpaController.java
@@ -8,33 +8,34 @@ package Persistencia;
 import Persistencia.exceptions.NonexistentEntityException;
 import Persistencia.exceptions.PreexistingEntityException;
 import java.io.Serializable;
+import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
 import model.Compra;
 
 /**
  *
  * @author daniel
  */
-public class CompraJpaController implements Serializable{
-    
-   //constructor
+public class CompraJpaController implements Serializable {
+
     public CompraJpaController(EntityManagerFactory emf) {
-    this.emf = emf;
-}
-    private EntityManagerFactory emf = null;
-    
-    public CompraJpaController() {
-   emf= Persistence.createEntityManagerFactory("ElMercaditoPU");
+        this.emf = emf;
     }
-    
+    private EntityManagerFactory emf = null;
+
+    CompraJpaController() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
-//metodo de jpaController. Create
     public void create(Compra compra) throws PreexistingEntityException, Exception {
         EntityManager em = null;
         try {
@@ -44,30 +45,29 @@ public class CompraJpaController implements Serializable{
             em.getTransaction().commit();
         } catch (Exception ex) {
             if (findCompra(compra.getNroCompra()) != null) {
-                throw new PreexistingEntityException("compra " + compra + " already exists.", ex);
+                throw new PreexistingEntityException("Compra " + compra + " already exists.", ex);
             }
             throw ex;
         } finally {
             if (em != null) {
                 em.close();
             }
-        
         }
     }
-//    metodo de jpaController. edit
+
     public void edit(Compra compra) throws NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            compra= em.merge(compra);
+            compra = em.merge(compra);
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
                 int id = compra.getNroCompra();
                 if (findCompra(id) == null) {
-                    throw new NonexistentEntityException("The compra with id " + id+ " no longer exists.");
+                    throw new NonexistentEntityException("The compra with id " + id + " no longer exists.");
                 }
             }
             throw ex;
@@ -77,16 +77,6 @@ public class CompraJpaController implements Serializable{
             }
         }
     }
-//    método de jpaController: obtener un elemento
-    public Compra findCompra(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Compra.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
 
     public void destroy(int id) throws NonexistentEntityException {
         EntityManager em = null;
@@ -98,7 +88,7 @@ public class CompraJpaController implements Serializable{
                 compra = em.getReference(Compra.class, id);
                 compra.getNroCompra();
             } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The compra with nroCliente " + id + " no longer exists.", enfe);
+                throw new NonexistentEntityException("The compra with id " + id + " no longer exists.", enfe);
             }
             em.remove(compra);
             em.getTransaction().commit();
@@ -108,4 +98,51 @@ public class CompraJpaController implements Serializable{
             }
         }
     }
+
+    public List<Compra> findCompraEntities() {
+        return findCompraEntities(true, -1, -1);
+    }
+
+    public List<Compra> findCompraEntities(int maxResults, int firstResult) {
+        return findCompraEntities(false, maxResults, firstResult);
+    }
+
+    private List<Compra> findCompraEntities(boolean all, int maxResults, int firstResult) {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            cq.select(cq.from(Compra.class));
+            Query q = em.createQuery(cq);
+            if (!all) {
+                q.setMaxResults(maxResults);
+                q.setFirstResult(firstResult);
+            }
+            return q.getResultList();
+        } finally {
+            em.close();
+        }
+    }
+
+    public Compra findCompra(int id) {
+        EntityManager em = getEntityManager();
+        try {
+            return em.find(Compra.class, id);
+        } finally {
+            em.close();
+        }
+    }
+
+    public int getCompraCount() {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            Root<Compra> rt = cq.from(Compra.class);
+            cq.select(em.getCriteriaBuilder().count(rt));
+            Query q = em.createQuery(cq);
+            return ((Long) q.getSingleResult()).intValue();
+        } finally {
+            em.close();
+        }
+    }
+    
 }
