--- src/Persistencia/CompraJpaController1.java
@@ -1,144 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Persistencia;
-
-import Persistencia.exceptions.NonexistentEntityException;
-import Persistencia.exceptions.PreexistingEntityException;
-import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
-import model.Compra;
-
-/**
- *
- * @author daniel
- */
-public class CompraJpaController1 implements Serializable {
-
-    public CompraJpaController1(EntityManagerFactory emf) {
-        this.emf = emf;
-    }
-    private EntityManagerFactory emf = null;
-
-    public EntityManager getEntityManager() {
-        return emf.createEntityManager();
-    }
-
-    public void create(Compra compra) throws PreexistingEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            em.persist(compra);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findCompra(compra.getNroCompra()) != null) {
-                throw new PreexistingEntityException("Compra " + compra + " already exists.", ex);
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void edit(Compra compra) throws NonexistentEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            compra = em.merge(compra);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            String msg = ex.getLocalizedMessage();
-            if (msg == null || msg.length() == 0) {
-                int id = compra.getNroCompra();
-                if (findCompra(id) == null) {
-                    throw new NonexistentEntityException("The compra with id " + id + " no longer exists.");
-                }
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void destroy(int id) throws NonexistentEntityException {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            Compra compra;
-            try {
-                compra = em.getReference(Compra.class, id);
-                compra.getNroCompra();
-            } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The compra with id " + id + " no longer exists.", enfe);
-            }
-            em.remove(compra);
-            em.getTransaction().commit();
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public List<Compra> findCompraEntities() {
-        return findCompraEntities(true, -1, -1);
-    }
-
-    public List<Compra> findCompraEntities(int maxResults, int firstResult) {
-        return findCompraEntities(false, maxResults, firstResult);
-    }
-
-    private List<Compra> findCompraEntities(boolean all, int maxResults, int firstResult) {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            cq.select(cq.from(Compra.class));
-            Query q = em.createQuery(cq);
-            if (!all) {
-                q.setMaxResults(maxResults);
-                q.setFirstResult(firstResult);
-            }
-            return q.getResultList();
-        } finally {
-            em.close();
-        }
-    }
-
-    public Compra findCompra(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Compra.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
-    public int getCompraCount() {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            Root<Compra> rt = cq.from(Compra.class);
-            cq.select(em.getCriteriaBuilder().count(rt));
-            Query q = em.createQuery(cq);
-            return ((Long) q.getSingleResult()).intValue();
-        } finally {
-            em.close();
-        }
-    }
-    
-}
