+++ src/Persistencia/ControladoraPersistencia.java
@@ -54,14 +54,15 @@ import model.VentaMinorista;
 import model.Zona;
 import Persistencia.exceptions.NonexistentEntityException;
 import Persistencia.exceptions.PreexistingEntityException;
+import java.util.LinkedList;
 import java.util.List;
 // * @author daniel
 
 public class ControladoraPersistencia {
     public ControladoraPersistencia() {
-        this.articuloPersistencia = new ArticuloJpaController1();
+        this.articuloPersistencia = new ArticuloJpaController();
     }
-    ArticuloJpaController1 articuloPersistencia;
+    ArticuloJpaController articuloPersistencia;
     BultoJpaController bultoPersistencia=new BultoJpaController();
     CajaJpaController cajaPersistencia = new CajaJpaController();
     CajeroJpaController cajeroPersistencia = new CajeroJpaController();
@@ -116,8 +117,8 @@ public class ControladoraPersistencia {
     public void BorrarArticuloPersis(int id) throws NonexistentEntityException{
         this.articuloPersistencia.destroy(id);
     }
-    public ArrayList BuscarListaArticulos(){
-       return (ArrayList) this.articuloPersistencia.findArticuloEntities();
+    public List<Articulo> BuscarListaArticulos(){
+       return this.articuloPersistencia.findArticuloEntities();
     }
     /*public Articulo BuscarUnArticulo(String id){
         return this.articuloPersistencia.findArticulo(id);
@@ -132,7 +133,7 @@ public class ControladoraPersistencia {
     public void BorrarBultoPersis(String id) throws NonexistentEntityException{
         this.bultoPersistencia.destroy(id);
     }
-    public List BuscarListaBultos(){
+    public List<Bulto> BuscarListaBultos(){
        return this.bultoPersistencia.findBultoEntities();
     }
     public Bulto BuscarUnBulto(String id){
@@ -150,7 +151,7 @@ public class ControladoraPersistencia {
     public void BorrarCajaPersis(int id) throws NonexistentEntityException{
         this.cajaPersistencia.destroy(id);
     }
-    public List BuscarListaCajas(){
+    public List<Caja> BuscarListaCajas(){
        return this.cajaPersistencia.findCajaEntities();
     }
     public Caja BuscarUnaCaja(int id){
@@ -168,7 +169,7 @@ public class ControladoraPersistencia {
     public void BorrarCajeroPersis(int id) throws NonexistentEntityException{
         this.cajeroPersistencia.destroy(id);
     }
-    public List BuscarListaCajeros(){
+    public List<Cajero> BuscarListaCajeros(){
        return this.cajeroPersistencia.findCajeroEntities();
     }
     public Cajero BuscarUnCajero(String id){
@@ -185,7 +186,7 @@ public class ControladoraPersistencia {
     public void BorrarClientePersis(int id) throws NonexistentEntityException{
         this.clientePersistencia.destroy(id);
     }
-    public List BuscarListaClientes(){
+    public List<Cliente> BuscarListaClientes(){
        return this.clientePersistencia.findClienteEntities();
     }
     public Object BuscarUnCliente(String id){
@@ -204,7 +205,7 @@ public class ControladoraPersistencia {
     public void BorrarProveedorPersis(int id) throws NonexistentEntityException{
         this.proveedorPersistencia.destroy(id);
     }
-    public List BuscarListaProveedores(){
+    public List<Proveedor> BuscarListaProveedores(){
        return this.proveedorPersistencia.findProveedorEntities();
     }
     public Object BuscarUnProveedor(String id){
@@ -221,7 +222,7 @@ public class ControladoraPersistencia {
     public void BorrarPersonaPersis(int id) throws NonexistentEntityException{
         this.personaPersistencia.destroy(id);
     }
-    public List BuscarListapersonas(){
+    public List<Persona> BuscarListaPersonas(){
        return this.personaPersistencia.findPersonaEntities();
     }
     public Object BuscarUnPersona(String id){
