+++ src/Persistencia/DescripcionArticuloJpaController.java
@@ -8,33 +8,34 @@ package Persistencia;
 import Persistencia.exceptions.NonexistentEntityException;
 import Persistencia.exceptions.PreexistingEntityException;
 import java.io.Serializable;
+import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
 import model.DescripcionArticulo;
 
 /**
  *
  * @author daniel
  */
-public class DescripcionArticuloJpaController implements Serializable{
-   
-   //constructor
+public class DescripcionArticuloJpaController implements Serializable {
+
     public DescripcionArticuloJpaController(EntityManagerFactory emf) {
-    this.emf = emf;
-}
-    private EntityManagerFactory emf = null;
-    
-    public DescripcionArticuloJpaController() {
-   emf= Persistence.createEntityManagerFactory("ElMercaditoPU");
+        this.emf = emf;
     }
-    
+    private EntityManagerFactory emf = null;
+
+    DescripcionArticuloJpaController() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
-//metodo de jpaController. Create
     public void create(DescripcionArticulo descripcionArticulo) throws PreexistingEntityException, Exception {
         EntityManager em = null;
         try {
@@ -44,30 +45,29 @@ public class DescripcionArticuloJpaController implements Serializable{
             em.getTransaction().commit();
         } catch (Exception ex) {
             if (findDescripcionArticulo(descripcionArticulo.getCodigoBarra()) != null) {
-                throw new PreexistingEntityException("descripcionArticulo " + descripcionArticulo + " already exists.", ex);
+                throw new PreexistingEntityException("DescripcionArticulo " + descripcionArticulo + " already exists.", ex);
             }
             throw ex;
         } finally {
             if (em != null) {
                 em.close();
             }
-        
         }
     }
-//    metodo de jpaController. edit
+
     public void edit(DescripcionArticulo descripcionArticulo) throws NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            descripcionArticulo= em.merge(descripcionArticulo);
+            descripcionArticulo = em.merge(descripcionArticulo);
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
                 String id = descripcionArticulo.getCodigoBarra();
                 if (findDescripcionArticulo(id) == null) {
-                    throw new NonexistentEntityException("The descripcionArticulo with id " + id+ " no longer exists.");
+                    throw new NonexistentEntityException("The descripcionArticulo with id " + id + " no longer exists.");
                 }
             }
             throw ex;
@@ -77,18 +77,8 @@ public class DescripcionArticuloJpaController implements Serializable{
             }
         }
     }
-//    método de jpaController: obtener un elemento
-    public DescripcionArticulo findDescripcionArticulo(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(DescripcionArticulo.class, id);
-        } finally {
-            em.close();
-        }
-    }
 
-
-    public void destroy(int id) throws NonexistentEntityException {
+    public void destroy(String id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -98,7 +88,7 @@ public class DescripcionArticuloJpaController implements Serializable{
                 descripcionArticulo = em.getReference(DescripcionArticulo.class, id);
                 descripcionArticulo.getCodigoBarra();
             } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The descripcionArticulo with Codigo Barra " + id + " no longer exists.", enfe);
+                throw new NonexistentEntityException("The descripcionArticulo with id " + id + " no longer exists.", enfe);
             }
             em.remove(descripcionArticulo);
             em.getTransaction().commit();
@@ -109,8 +99,50 @@ public class DescripcionArticuloJpaController implements Serializable{
         }
     }
 
-    private Object findDescripcionArticulo(String id) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    public List<DescripcionArticulo> findDescripcionArticuloEntities() {
+        return findDescripcionArticuloEntities(true, -1, -1);
     }
-}
 
+    public List<DescripcionArticulo> findDescripcionArticuloEntities(int maxResults, int firstResult) {
+        return findDescripcionArticuloEntities(false, maxResults, firstResult);
+    }
+
+    private List<DescripcionArticulo> findDescripcionArticuloEntities(boolean all, int maxResults, int firstResult) {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            cq.select(cq.from(DescripcionArticulo.class));
+            Query q = em.createQuery(cq);
+            if (!all) {
+                q.setMaxResults(maxResults);
+                q.setFirstResult(firstResult);
+            }
+            return q.getResultList();
+        } finally {
+            em.close();
+        }
+    }
+
+    public DescripcionArticulo findDescripcionArticulo(String id) {
+        EntityManager em = getEntityManager();
+        try {
+            return em.find(DescripcionArticulo.class, id);
+        } finally {
+            em.close();
+        }
+    }
+
+    public int getDescripcionArticuloCount() {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            Root<DescripcionArticulo> rt = cq.from(DescripcionArticulo.class);
+            cq.select(em.getCriteriaBuilder().count(rt));
+            Query q = em.createQuery(cq);
+            return ((Long) q.getSingleResult()).intValue();
+        } finally {
+            em.close();
+        }
+    }
+    
+}
