--- src/Persistencia/DireccionJpaController1.java
@@ -1,138 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Persistencia;
-
-import Persistencia.exceptions.NonexistentEntityException;
-import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
-import model.Direccion;
-
-/**
- *
- * @author daniel
- */
-public class DireccionJpaController1 implements Serializable {
-
-    public DireccionJpaController1(EntityManagerFactory emf) {
-        this.emf = emf;
-    }
-    private EntityManagerFactory emf = null;
-
-    public EntityManager getEntityManager() {
-        return emf.createEntityManager();
-    }
-
-    public void create(Direccion direccion) {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            em.persist(direccion);
-            em.getTransaction().commit();
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void edit(Direccion direccion) throws NonexistentEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            direccion = em.merge(direccion);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            String msg = ex.getLocalizedMessage();
-            if (msg == null || msg.length() == 0) {
-                int id = direccion.getCodigo();
-                if (findDireccion(id) == null) {
-                    throw new NonexistentEntityException("The direccion with id " + id + " no longer exists.");
-                }
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void destroy(int id) throws NonexistentEntityException {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            Direccion direccion;
-            try {
-                direccion = em.getReference(Direccion.class, id);
-                direccion.getCodigo();
-            } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The direccion with id " + id + " no longer exists.", enfe);
-            }
-            em.remove(direccion);
-            em.getTransaction().commit();
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public List<Direccion> findDireccionEntities() {
-        return findDireccionEntities(true, -1, -1);
-    }
-
-    public List<Direccion> findDireccionEntities(int maxResults, int firstResult) {
-        return findDireccionEntities(false, maxResults, firstResult);
-    }
-
-    private List<Direccion> findDireccionEntities(boolean all, int maxResults, int firstResult) {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            cq.select(cq.from(Direccion.class));
-            Query q = em.createQuery(cq);
-            if (!all) {
-                q.setMaxResults(maxResults);
-                q.setFirstResult(firstResult);
-            }
-            return q.getResultList();
-        } finally {
-            em.close();
-        }
-    }
-
-    public Direccion findDireccion(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Direccion.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
-    public int getDireccionCount() {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            Root<Direccion> rt = cq.from(Direccion.class);
-            cq.select(em.getCriteriaBuilder().count(rt));
-            Query q = em.createQuery(cq);
-            return ((Long) q.getSingleResult()).intValue();
-        } finally {
-            em.close();
-        }
-    }
-    
-}
