+++ src/Persistencia/EmpleadoJpaController.java
@@ -8,34 +8,34 @@ package Persistencia;
 import Persistencia.exceptions.NonexistentEntityException;
 import Persistencia.exceptions.PreexistingEntityException;
 import java.io.Serializable;
+import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
 import model.Empleado;
 
 /**
  *
  * @author daniel
  */
+public class EmpleadoJpaController implements Serializable {
 
-public class EmpleadoJpaController implements Serializable{
-    
-   //constructor
     public EmpleadoJpaController(EntityManagerFactory emf) {
-    this.emf = emf;
-}
-    private EntityManagerFactory emf = null;
-    
-    public EmpleadoJpaController() {
-   emf= Persistence.createEntityManagerFactory("ElMercaditoPU");
+        this.emf = emf;
     }
-    
+    private EntityManagerFactory emf = null;
+
+    EmpleadoJpaController() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
-//metodo de jpaController. Create
     public void create(Empleado empleado) throws PreexistingEntityException, Exception {
         EntityManager em = null;
         try {
@@ -52,23 +52,22 @@ public class EmpleadoJpaController implements Serializable{
             if (em != null) {
                 em.close();
             }
-        
         }
     }
-//    metodo de jpaController. edit
+
     public void edit(Empleado empleado) throws NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            empleado= em.merge(empleado);
+            empleado = em.merge(empleado);
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
                 String id = empleado.getDni();
                 if (findEmpleado(id) == null) {
-                    throw new NonexistentEntityException("The empleado with id " + id+ " no longer exists.");
+                    throw new NonexistentEntityException("The empleado with id " + id + " no longer exists.");
                 }
             }
             throw ex;
@@ -78,18 +77,8 @@ public class EmpleadoJpaController implements Serializable{
             }
         }
     }
-//    método de jpaController: obtener un elemento
-    public Empleado findEmpleado(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Empleado.class, id);
-        } finally {
-            em.close();
-        }
-    }
 
-
-    public void destroy(int id) throws NonexistentEntityException {
+    public void destroy(String id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -99,7 +88,7 @@ public class EmpleadoJpaController implements Serializable{
                 empleado = em.getReference(Empleado.class, id);
                 empleado.getDni();
             } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The empleado with dni " + id + " no longer exists.", enfe);
+                throw new NonexistentEntityException("The empleado with id " + id + " no longer exists.", enfe);
             }
             em.remove(empleado);
             em.getTransaction().commit();
@@ -110,7 +99,50 @@ public class EmpleadoJpaController implements Serializable{
         }
     }
 
-    private Object findEmpleado(String id) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    public List<Empleado> findEmpleadoEntities() {
+        return findEmpleadoEntities(true, -1, -1);
     }
+
+    public List<Empleado> findEmpleadoEntities(int maxResults, int firstResult) {
+        return findEmpleadoEntities(false, maxResults, firstResult);
+    }
+
+    private List<Empleado> findEmpleadoEntities(boolean all, int maxResults, int firstResult) {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            cq.select(cq.from(Empleado.class));
+            Query q = em.createQuery(cq);
+            if (!all) {
+                q.setMaxResults(maxResults);
+                q.setFirstResult(firstResult);
+            }
+            return q.getResultList();
+        } finally {
+            em.close();
+        }
+    }
+
+    public Empleado findEmpleado(String id) {
+        EntityManager em = getEntityManager();
+        try {
+            return em.find(Empleado.class, id);
+        } finally {
+            em.close();
+        }
+    }
+
+    public int getEmpleadoCount() {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            Root<Empleado> rt = cq.from(Empleado.class);
+            cq.select(em.getCriteriaBuilder().count(rt));
+            Query q = em.createQuery(cq);
+            return ((Long) q.getSingleResult()).intValue();
+        } finally {
+            em.close();
+        }
+    }
+    
 }
