+++ src/Persistencia/EncargadoJpaController.java
@@ -8,34 +8,34 @@ package Persistencia;
 import Persistencia.exceptions.NonexistentEntityException;
 import Persistencia.exceptions.PreexistingEntityException;
 import java.io.Serializable;
+import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
 import model.Encargado;
 
 /**
  *
  * @author daniel
  */
+public class EncargadoJpaController implements Serializable {
 
-public class EncargadoJpaController implements Serializable{
-    
-   //constructor
     public EncargadoJpaController(EntityManagerFactory emf) {
-    this.emf = emf;
-}
-    private EntityManagerFactory emf = null;
-    
-    public EncargadoJpaController() {
-   emf= Persistence.createEntityManagerFactory("ElMercaditoPU");
+        this.emf = emf;
     }
-    
+    private EntityManagerFactory emf = null;
+
+    EncargadoJpaController() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
-//metodo de jpaController. Create
     public void create(Encargado encargado) throws PreexistingEntityException, Exception {
         EntityManager em = null;
         try {
@@ -52,23 +52,22 @@ public class EncargadoJpaController implements Serializable{
             if (em != null) {
                 em.close();
             }
-        
         }
     }
-//    metodo de jpaController. edit
+
     public void edit(Encargado encargado) throws NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            encargado= em.merge(encargado);
+            encargado = em.merge(encargado);
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
                 String id = encargado.getDni();
                 if (findEncargado(id) == null) {
-                    throw new NonexistentEntityException("The encargado with id " + id+ " no longer exists.");
+                    throw new NonexistentEntityException("The encargado with id " + id + " no longer exists.");
                 }
             }
             throw ex;
@@ -78,28 +77,18 @@ public class EncargadoJpaController implements Serializable{
             }
         }
     }
-//    método de jpaController: obtener un elemento
-    public Encargado findEncargado(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Encargado.class, id);
-        } finally {
-            em.close();
-        }
-    }
 
-
-    public void destroy(int id) throws NonexistentEntityException {
+    public void destroy(String id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
             Encargado encargado;
             try {
-               encargado = em.getReference(Encargado.class, id);
+                encargado = em.getReference(Encargado.class, id);
                 encargado.getDni();
             } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The encargado with dni " + id + " no longer exists.", enfe);
+                throw new NonexistentEntityException("The encargado with id " + id + " no longer exists.", enfe);
             }
             em.remove(encargado);
             em.getTransaction().commit();
@@ -110,9 +99,50 @@ public class EncargadoJpaController implements Serializable{
         }
     }
 
-    private Object findEncargado(String id) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    public List<Encargado> findEncargadoEntities() {
+        return findEncargadoEntities(true, -1, -1);
     }
+
+    public List<Encargado> findEncargadoEntities(int maxResults, int firstResult) {
+        return findEncargadoEntities(false, maxResults, firstResult);
+    }
+
+    private List<Encargado> findEncargadoEntities(boolean all, int maxResults, int firstResult) {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            cq.select(cq.from(Encargado.class));
+            Query q = em.createQuery(cq);
+            if (!all) {
+                q.setMaxResults(maxResults);
+                q.setFirstResult(firstResult);
+            }
+            return q.getResultList();
+        } finally {
+            em.close();
+        }
+    }
+
+    public Encargado findEncargado(String id) {
+        EntityManager em = getEntityManager();
+        try {
+            return em.find(Encargado.class, id);
+        } finally {
+            em.close();
+        }
+    }
+
+    public int getEncargadoCount() {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            Root<Encargado> rt = cq.from(Encargado.class);
+            cq.select(em.getCriteriaBuilder().count(rt));
+            Query q = em.createQuery(cq);
+            return ((Long) q.getSingleResult()).intValue();
+        } finally {
+            em.close();
+        }
+    }
+    
 }
-
-
