--- src/Persistencia/GondolaJpaController1.java
@@ -1,144 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Persistencia;
-
-import Persistencia.exceptions.NonexistentEntityException;
-import Persistencia.exceptions.PreexistingEntityException;
-import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
-import model.Gondola;
-
-/**
- *
- * @author daniel
- */
-public class GondolaJpaController1 implements Serializable {
-
-    public GondolaJpaController1(EntityManagerFactory emf) {
-        this.emf = emf;
-    }
-    private EntityManagerFactory emf = null;
-
-    public EntityManager getEntityManager() {
-        return emf.createEntityManager();
-    }
-
-    public void create(Gondola gondola) throws PreexistingEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            em.persist(gondola);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findGondola(gondola.getNroGondola()) != null) {
-                throw new PreexistingEntityException("Gondola " + gondola + " already exists.", ex);
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void edit(Gondola gondola) throws NonexistentEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            gondola = em.merge(gondola);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            String msg = ex.getLocalizedMessage();
-            if (msg == null || msg.length() == 0) {
-                int id = gondola.getNroGondola();
-                if (findGondola(id) == null) {
-                    throw new NonexistentEntityException("The gondola with id " + id + " no longer exists.");
-                }
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void destroy(int id) throws NonexistentEntityException {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            Gondola gondola;
-            try {
-                gondola = em.getReference(Gondola.class, id);
-                gondola.getNroGondola();
-            } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The gondola with id " + id + " no longer exists.", enfe);
-            }
-            em.remove(gondola);
-            em.getTransaction().commit();
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public List<Gondola> findGondolaEntities() {
-        return findGondolaEntities(true, -1, -1);
-    }
-
-    public List<Gondola> findGondolaEntities(int maxResults, int firstResult) {
-        return findGondolaEntities(false, maxResults, firstResult);
-    }
-
-    private List<Gondola> findGondolaEntities(boolean all, int maxResults, int firstResult) {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            cq.select(cq.from(Gondola.class));
-            Query q = em.createQuery(cq);
-            if (!all) {
-                q.setMaxResults(maxResults);
-                q.setFirstResult(firstResult);
-            }
-            return q.getResultList();
-        } finally {
-            em.close();
-        }
-    }
-
-    public Gondola findGondola(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Gondola.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
-    public int getGondolaCount() {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            Root<Gondola> rt = cq.from(Gondola.class);
-            cq.select(em.getCriteriaBuilder().count(rt));
-            Query q = em.createQuery(cq);
-            return ((Long) q.getSingleResult()).intValue();
-        } finally {
-            em.close();
-        }
-    }
-    
-}
