+++ src/Persistencia/PagoJpaController.java
@@ -6,68 +6,62 @@
 package Persistencia;
 
 import Persistencia.exceptions.NonexistentEntityException;
-import Persistencia.exceptions.PreexistingEntityException;
 import java.io.Serializable;
+import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
 import model.Pago;
 
 /**
  *
  * @author daniel
  */
+public class PagoJpaController implements Serializable {
 
-public class PagoJpaController implements Serializable{
-    
-   //constructor
     public PagoJpaController(EntityManagerFactory emf) {
-    this.emf = emf;
-}
-    private EntityManagerFactory emf = null;
-    public PagoJpaController() {
-   emf= Persistence.createEntityManagerFactory("ElMercaditoPU");
+        this.emf = emf;
     }
-    
+    private EntityManagerFactory emf = null;
+
+    PagoJpaController() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
-//metodo de jpaController. Create
-    public void create(Pago pago) throws PreexistingEntityException, Exception {
+    public void create(Pago pago) {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
             em.persist(pago);
             em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findPago(pago.getCodigoPago()) != null) {
-                throw new PreexistingEntityException("Pago " + pago + " already exists.", ex);
-            }
-            throw ex;
         } finally {
             if (em != null) {
                 em.close();
             }
-        
         }
     }
-//    metodo de jpaController. edit
+
     public void edit(Pago pago) throws NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            pago= em.merge(pago);
+            pago = em.merge(pago);
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                String id = pago.getTipopago();
+                int id = pago.getCodigoPago();
                 if (findPago(id) == null) {
-                    throw new NonexistentEntityException("The pago with id " + id+ " no longer exists.");
+                    throw new NonexistentEntityException("The pago with id " + id + " no longer exists.");
                 }
             }
             throw ex;
@@ -77,16 +71,6 @@ public class PagoJpaController implements Serializable{
             }
         }
     }
-//    método de jpaController: obtener un elemento
-    public Pago findPago(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Pago.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
 
     public void destroy(int id) throws NonexistentEntityException {
         EntityManager em = null;
@@ -98,7 +82,7 @@ public class PagoJpaController implements Serializable{
                 pago = em.getReference(Pago.class, id);
                 pago.getCodigoPago();
             } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The pago with codigoPago " + id + " no longer exists.", enfe);
+                throw new NonexistentEntityException("The pago with id " + id + " no longer exists.", enfe);
             }
             em.remove(pago);
             em.getTransaction().commit();
@@ -109,7 +93,50 @@ public class PagoJpaController implements Serializable{
         }
     }
 
-    private Object findPago(String id) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    public List<Pago> findPagoEntities() {
+        return findPagoEntities(true, -1, -1);
     }
+
+    public List<Pago> findPagoEntities(int maxResults, int firstResult) {
+        return findPagoEntities(false, maxResults, firstResult);
+    }
+
+    private List<Pago> findPagoEntities(boolean all, int maxResults, int firstResult) {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            cq.select(cq.from(Pago.class));
+            Query q = em.createQuery(cq);
+            if (!all) {
+                q.setMaxResults(maxResults);
+                q.setFirstResult(firstResult);
+            }
+            return q.getResultList();
+        } finally {
+            em.close();
+        }
+    }
+
+    public Pago findPago(int id) {
+        EntityManager em = getEntityManager();
+        try {
+            return em.find(Pago.class, id);
+        } finally {
+            em.close();
+        }
+    }
+
+    public int getPagoCount() {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            Root<Pago> rt = cq.from(Pago.class);
+            cq.select(em.getCriteriaBuilder().count(rt));
+            Query q = em.createQuery(cq);
+            return ((Long) q.getSingleResult()).intValue();
+        } finally {
+            em.close();
+        }
+    }
+    
 }
