+++ src/Persistencia/PersonaJpaController.java
@@ -11,32 +11,31 @@ import java.io.Serializable;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
 import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
 import model.Persona;
 
 /**
  *
  * @author daniel
  */
+public class PersonaJpaController implements Serializable {
 
-public class PersonaJpaController implements Serializable{
-    
-   //constructor
     public PersonaJpaController(EntityManagerFactory emf) {
-    this.emf = emf;
-}
-    private EntityManagerFactory emf = null;
-    public PersonaJpaController() {
-   emf= Persistence.createEntityManagerFactory("ElMercaditoPU");
+        this.emf = emf;
     }
-    
+    private EntityManagerFactory emf = null;
+
+    PersonaJpaController() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
-//metodo de jpaController. Create
     public void create(Persona persona) throws PreexistingEntityException, Exception {
         EntityManager em = null;
         try {
@@ -53,23 +52,22 @@ public class PersonaJpaController implements Serializable{
             if (em != null) {
                 em.close();
             }
-        
         }
     }
-//    metodo de jpaController. edit
+
     public void edit(Persona persona) throws NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            persona= em.merge(persona);
+            persona = em.merge(persona);
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
                 String id = persona.getDni();
                 if (findPersona(id) == null) {
-                    throw new NonexistentEntityException("The persona with id " + id+ " no longer exists.");
+                    throw new NonexistentEntityException("The persona with id " + id + " no longer exists.");
                 }
             }
             throw ex;
@@ -79,16 +77,6 @@ public class PersonaJpaController implements Serializable{
             }
         }
     }
-//    método de jpaController: obtener un elemento
-    public Persona findPersona(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Persona.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
 
     public void destroy(int id) throws NonexistentEntityException {
         EntityManager em = null;
@@ -100,7 +88,7 @@ public class PersonaJpaController implements Serializable{
                 persona = em.getReference(Persona.class, id);
                 persona.getDni();
             } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The persona with dni " + id + " no longer exists.", enfe);
+                throw new NonexistentEntityException("The persona with id " + id + " no longer exists.", enfe);
             }
             em.remove(persona);
             em.getTransaction().commit();
@@ -110,7 +98,7 @@ public class PersonaJpaController implements Serializable{
             }
         }
     }
-    //método de jpa controller: obtener varios elementos
+
     public List<Persona> findPersonaEntities() {
         return findPersonaEntities(true, -1, -1);
     }
@@ -122,7 +110,9 @@ public class PersonaJpaController implements Serializable{
     private List<Persona> findPersonaEntities(boolean all, int maxResults, int firstResult) {
         EntityManager em = getEntityManager();
         try {
-            Query q = em.createQuery("select object(o) from Persona as o");
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            cq.select(cq.from(Persona.class));
+            Query q = em.createQuery(cq);
             if (!all) {
                 q.setMaxResults(maxResults);
                 q.setFirstResult(firstResult);
@@ -133,23 +123,26 @@ public class PersonaJpaController implements Serializable{
         }
     }
 
-
-//metodo de jpaController: obtener cantidad de elementos persistidos
-    public int getPersonaCount() {
+    public Persona findPersona(String id) {
         EntityManager em = getEntityManager();
         try {
-            Query q = em.createQuery("select count(o) from Persona as o");
-            return ((Long) q.getSingleResult()).intValue();
+            return em.find(Persona.class, id);
         } finally {
             em.close();
         }
     }
 
-        public Object findPersona(String id) {
-            throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    public int getPersonaCount() {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            Root<Persona> rt = cq.from(Persona.class);
+            cq.select(em.getCriteriaBuilder().count(rt));
+            Query q = em.createQuery(cq);
+            return ((Long) q.getSingleResult()).intValue();
+        } finally {
+            em.close();
         }
-
-   
     }
-
-
+    
+}
