+++ src/Persistencia/SeccionJpaController.java
@@ -8,33 +8,34 @@ package Persistencia;
 import Persistencia.exceptions.NonexistentEntityException;
 import Persistencia.exceptions.PreexistingEntityException;
 import java.io.Serializable;
+import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
 import model.Seccion;
 
 /**
  *
  * @author daniel
  */
-   
-public class SeccionJpaController implements Serializable{
-    
-   //constructor
+public class SeccionJpaController implements Serializable {
+
     public SeccionJpaController(EntityManagerFactory emf) {
-    this.emf = emf;
-}
-    private EntityManagerFactory emf = null;
-    public SeccionJpaController() {
-   emf= Persistence.createEntityManagerFactory("ElMercaditoPU");
+        this.emf = emf;
     }
-    
+    private EntityManagerFactory emf = null;
+
+    SeccionJpaController() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
-//metodo de jpaController. Create
     public void create(Seccion seccion) throws PreexistingEntityException, Exception {
         EntityManager em = null;
         try {
@@ -43,7 +44,7 @@ public class SeccionJpaController implements Serializable{
             em.persist(seccion);
             em.getTransaction().commit();
         } catch (Exception ex) {
-            if (null != findSeccion(seccion.getNroSeccion())) {
+            if (findSeccion(seccion.getNroSeccion()) != null) {
                 throw new PreexistingEntityException("Seccion " + seccion + " already exists.", ex);
             }
             throw ex;
@@ -51,23 +52,22 @@ public class SeccionJpaController implements Serializable{
             if (em != null) {
                 em.close();
             }
-        
         }
     }
-//    metodo de jpaController. edit
+
     public void edit(Seccion seccion) throws NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            seccion= em.merge(seccion);
+            seccion = em.merge(seccion);
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
                 int id = seccion.getNroSeccion();
                 if (findSeccion(id) == null) {
-                    throw new NonexistentEntityException("The seccion with id " + id+ " no longer exists.");
+                    throw new NonexistentEntityException("The seccion with id " + id + " no longer exists.");
                 }
             }
             throw ex;
@@ -77,16 +77,6 @@ public class SeccionJpaController implements Serializable{
             }
         }
     }
-//    método de jpaController: obtener un elemento
-    public Seccion findSeccion(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Seccion.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
 
     public void destroy(int id) throws NonexistentEntityException {
         EntityManager em = null;
@@ -98,7 +88,7 @@ public class SeccionJpaController implements Serializable{
                 seccion = em.getReference(Seccion.class, id);
                 seccion.getNroSeccion();
             } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The seccion with nroSeccion " + id + " no longer exists.", enfe);
+                throw new NonexistentEntityException("The seccion with id " + id + " no longer exists.", enfe);
             }
             em.remove(seccion);
             em.getTransaction().commit();
@@ -108,6 +98,51 @@ public class SeccionJpaController implements Serializable{
             }
         }
     }
-        
+
+    public List<Seccion> findSeccionEntities() {
+        return findSeccionEntities(true, -1, -1);
     }
 
+    public List<Seccion> findSeccionEntities(int maxResults, int firstResult) {
+        return findSeccionEntities(false, maxResults, firstResult);
+    }
+
+    private List<Seccion> findSeccionEntities(boolean all, int maxResults, int firstResult) {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            cq.select(cq.from(Seccion.class));
+            Query q = em.createQuery(cq);
+            if (!all) {
+                q.setMaxResults(maxResults);
+                q.setFirstResult(firstResult);
+            }
+            return q.getResultList();
+        } finally {
+            em.close();
+        }
+    }
+
+    public Seccion findSeccion(int id) {
+        EntityManager em = getEntityManager();
+        try {
+            return em.find(Seccion.class, id);
+        } finally {
+            em.close();
+        }
+    }
+
+    public int getSeccionCount() {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            Root<Seccion> rt = cq.from(Seccion.class);
+            cq.select(em.getCriteriaBuilder().count(rt));
+            Query q = em.createQuery(cq);
+            return ((Long) q.getSingleResult()).intValue();
+        } finally {
+            em.close();
+        }
+    }
+    
+}
