--- src/Persistencia/SeccionJpaController1.java
@@ -1,144 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Persistencia;
-
-import Persistencia.exceptions.NonexistentEntityException;
-import Persistencia.exceptions.PreexistingEntityException;
-import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
-import model.Seccion;
-
-/**
- *
- * @author daniel
- */
-public class SeccionJpaController1 implements Serializable {
-
-    public SeccionJpaController1(EntityManagerFactory emf) {
-        this.emf = emf;
-    }
-    private EntityManagerFactory emf = null;
-
-    public EntityManager getEntityManager() {
-        return emf.createEntityManager();
-    }
-
-    public void create(Seccion seccion) throws PreexistingEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            em.persist(seccion);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findSeccion(seccion.getNroSeccion()) != null) {
-                throw new PreexistingEntityException("Seccion " + seccion + " already exists.", ex);
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void edit(Seccion seccion) throws NonexistentEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            seccion = em.merge(seccion);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            String msg = ex.getLocalizedMessage();
-            if (msg == null || msg.length() == 0) {
-                int id = seccion.getNroSeccion();
-                if (findSeccion(id) == null) {
-                    throw new NonexistentEntityException("The seccion with id " + id + " no longer exists.");
-                }
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void destroy(int id) throws NonexistentEntityException {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            Seccion seccion;
-            try {
-                seccion = em.getReference(Seccion.class, id);
-                seccion.getNroSeccion();
-            } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The seccion with id " + id + " no longer exists.", enfe);
-            }
-            em.remove(seccion);
-            em.getTransaction().commit();
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public List<Seccion> findSeccionEntities() {
-        return findSeccionEntities(true, -1, -1);
-    }
-
-    public List<Seccion> findSeccionEntities(int maxResults, int firstResult) {
-        return findSeccionEntities(false, maxResults, firstResult);
-    }
-
-    private List<Seccion> findSeccionEntities(boolean all, int maxResults, int firstResult) {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            cq.select(cq.from(Seccion.class));
-            Query q = em.createQuery(cq);
-            if (!all) {
-                q.setMaxResults(maxResults);
-                q.setFirstResult(firstResult);
-            }
-            return q.getResultList();
-        } finally {
-            em.close();
-        }
-    }
-
-    public Seccion findSeccion(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Seccion.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
-    public int getSeccionCount() {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            Root<Seccion> rt = cq.from(Seccion.class);
-            cq.select(em.getCriteriaBuilder().count(rt));
-            Query q = em.createQuery(cq);
-            return ((Long) q.getSingleResult()).intValue();
-        } finally {
-            em.close();
-        }
-    }
-    
-}
