--- src/Persistencia/SucursalJpaController1.java
@@ -1,144 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Persistencia;
-
-import Persistencia.exceptions.NonexistentEntityException;
-import Persistencia.exceptions.PreexistingEntityException;
-import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
-import model.Sucursal;
-
-/**
- *
- * @author daniel
- */
-public class SucursalJpaController1 implements Serializable {
-
-    public SucursalJpaController1(EntityManagerFactory emf) {
-        this.emf = emf;
-    }
-    private EntityManagerFactory emf = null;
-
-    public EntityManager getEntityManager() {
-        return emf.createEntityManager();
-    }
-
-    public void create(Sucursal sucursal) throws PreexistingEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            em.persist(sucursal);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findSucursal(sucursal.getNroSucursal()) != null) {
-                throw new PreexistingEntityException("Sucursal " + sucursal + " already exists.", ex);
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void edit(Sucursal sucursal) throws NonexistentEntityException, Exception {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            sucursal = em.merge(sucursal);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            String msg = ex.getLocalizedMessage();
-            if (msg == null || msg.length() == 0) {
-                int id = sucursal.getNroSucursal();
-                if (findSucursal(id) == null) {
-                    throw new NonexistentEntityException("The sucursal with id " + id + " no longer exists.");
-                }
-            }
-            throw ex;
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public void destroy(int id) throws NonexistentEntityException {
-        EntityManager em = null;
-        try {
-            em = getEntityManager();
-            em.getTransaction().begin();
-            Sucursal sucursal;
-            try {
-                sucursal = em.getReference(Sucursal.class, id);
-                sucursal.getNroSucursal();
-            } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The sucursal with id " + id + " no longer exists.", enfe);
-            }
-            em.remove(sucursal);
-            em.getTransaction().commit();
-        } finally {
-            if (em != null) {
-                em.close();
-            }
-        }
-    }
-
-    public List<Sucursal> findSucursalEntities() {
-        return findSucursalEntities(true, -1, -1);
-    }
-
-    public List<Sucursal> findSucursalEntities(int maxResults, int firstResult) {
-        return findSucursalEntities(false, maxResults, firstResult);
-    }
-
-    private List<Sucursal> findSucursalEntities(boolean all, int maxResults, int firstResult) {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            cq.select(cq.from(Sucursal.class));
-            Query q = em.createQuery(cq);
-            if (!all) {
-                q.setMaxResults(maxResults);
-                q.setFirstResult(firstResult);
-            }
-            return q.getResultList();
-        } finally {
-            em.close();
-        }
-    }
-
-    public Sucursal findSucursal(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Sucursal.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
-    public int getSucursalCount() {
-        EntityManager em = getEntityManager();
-        try {
-            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
-            Root<Sucursal> rt = cq.from(Sucursal.class);
-            cq.select(em.getCriteriaBuilder().count(rt));
-            Query q = em.createQuery(cq);
-            return ((Long) q.getSingleResult()).intValue();
-        } finally {
-            em.close();
-        }
-    }
-    
-}
