+++ src/Persistencia/ZonaJpaController.java
@@ -6,69 +6,109 @@
 package Persistencia;
 
 import Persistencia.exceptions.NonexistentEntityException;
-import Persistencia.exceptions.PreexistingEntityException;
 import java.io.Serializable;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+import model.Provincia;
+import java.util.ArrayList;
+import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
 import model.Zona;
 
 /**
  *
  * @author daniel
  */
-public class ZonaJpaController implements Serializable{
-  
-     
-   //constructor
+public class ZonaJpaController implements Serializable {
+
     public ZonaJpaController(EntityManagerFactory emf) {
-    this.emf = emf;
-}
-    private EntityManagerFactory emf = null;
-    
-    public ZonaJpaController() {
-   emf= Persistence.createEntityManagerFactory("ElMercaditoPU");
+        this.emf = emf;
     }
-    
+    private EntityManagerFactory emf = null;
+
+    ZonaJpaController() {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
-//metodo de jpaController. Create
-    public void create(Zona zona) throws PreexistingEntityException, Exception {
+    public void create(Zona zona) {
+        if (zona.getListaProvincias() == null) {
+            zona.setListaProvincias(new ArrayList<Provincia>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            em.persist(zona);
-            em.getTransaction().commit();
-        } catch (Exception ex) {
-            if (findZona(zona.getCodigoZona()) != null) {
-                throw new PreexistingEntityException("Zona" + zona + " already exists.", ex);
+            ArrayList<Provincia> attachedListaProvincias = new ArrayList<Provincia>();
+            for (Provincia listaProvinciasProvinciaToAttach : zona.getListaProvincias()) {
+                listaProvinciasProvinciaToAttach = em.getReference(listaProvinciasProvinciaToAttach.getClass(), listaProvinciasProvinciaToAttach.getCodigoProv());
+                attachedListaProvincias.add(listaProvinciasProvinciaToAttach);
             }
-            throw ex;
+            zona.setListaProvincias(attachedListaProvincias);
+            em.persist(zona);
+            for (Provincia listaProvinciasProvincia : zona.getListaProvincias()) {
+                Zona oldUnaZonaOfListaProvinciasProvincia = listaProvinciasProvincia.getUnaZona();
+                listaProvinciasProvincia.setUnaZona(zona);
+                listaProvinciasProvincia = em.merge(listaProvinciasProvincia);
+                if (oldUnaZonaOfListaProvinciasProvincia != null) {
+                    oldUnaZonaOfListaProvinciasProvincia.getListaProvincias().remove(listaProvinciasProvincia);
+                    oldUnaZonaOfListaProvinciasProvincia = em.merge(oldUnaZonaOfListaProvinciasProvincia);
+                }
+            }
+            em.getTransaction().commit();
         } finally {
             if (em != null) {
                 em.close();
             }
-        
         }
     }
-//    metodo de jpaController. edit
+
     public void edit(Zona zona) throws NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
-            zona= em.merge(zona);
+            Zona persistentZona = em.find(Zona.class, zona.getCodigoZona());
+            ArrayList<Provincia> listaProvinciasOld = persistentZona.getListaProvincias();
+            ArrayList<Provincia> listaProvinciasNew = zona.getListaProvincias();
+            ArrayList<Provincia> attachedListaProvinciasNew = new ArrayList<Provincia>();
+            for (Provincia listaProvinciasNewProvinciaToAttach : listaProvinciasNew) {
+                listaProvinciasNewProvinciaToAttach = em.getReference(listaProvinciasNewProvinciaToAttach.getClass(), listaProvinciasNewProvinciaToAttach.getCodigoProv());
+                attachedListaProvinciasNew.add(listaProvinciasNewProvinciaToAttach);
+            }
+            listaProvinciasNew = attachedListaProvinciasNew;
+            zona.setListaProvincias(listaProvinciasNew);
+            zona = em.merge(zona);
+            for (Provincia listaProvinciasOldProvincia : listaProvinciasOld) {
+                if (!listaProvinciasNew.contains(listaProvinciasOldProvincia)) {
+                    listaProvinciasOldProvincia.setUnaZona(null);
+                    listaProvinciasOldProvincia = em.merge(listaProvinciasOldProvincia);
+                }
+            }
+            for (Provincia listaProvinciasNewProvincia : listaProvinciasNew) {
+                if (!listaProvinciasOld.contains(listaProvinciasNewProvincia)) {
+                    Zona oldUnaZonaOfListaProvinciasNewProvincia = listaProvinciasNewProvincia.getUnaZona();
+                    listaProvinciasNewProvincia.setUnaZona(zona);
+                    listaProvinciasNewProvincia = em.merge(listaProvinciasNewProvincia);
+                    if (oldUnaZonaOfListaProvinciasNewProvincia != null && !oldUnaZonaOfListaProvinciasNewProvincia.equals(zona)) {
+                        oldUnaZonaOfListaProvinciasNewProvincia.getListaProvincias().remove(listaProvinciasNewProvincia);
+                        oldUnaZonaOfListaProvinciasNewProvincia = em.merge(oldUnaZonaOfListaProvinciasNewProvincia);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
                 int id = zona.getCodigoZona();
                 if (findZona(id) == null) {
-                    throw new NonexistentEntityException("The Zona with id " + id+ " no longer exists.");
+                    throw new NonexistentEntityException("The zona with id " + id + " no longer exists.");
                 }
             }
             throw ex;
@@ -78,16 +118,6 @@ public class ZonaJpaController implements Serializable{
             }
         }
     }
-//    método de jpaController: obtener un elemento
-    public Zona findZona(int id) {
-        EntityManager em = getEntityManager();
-        try {
-            return em.find(Zona.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
 
     public void destroy(int id) throws NonexistentEntityException {
         EntityManager em = null;
@@ -99,7 +129,12 @@ public class ZonaJpaController implements Serializable{
                 zona = em.getReference(Zona.class, id);
                 zona.getCodigoZona();
             } catch (EntityNotFoundException enfe) {
-                throw new NonexistentEntityException("The zona with codigoZona " + id + " no longer exists.", enfe);
+                throw new NonexistentEntityException("The zona with id " + id + " no longer exists.", enfe);
+            }
+            ArrayList<Provincia> listaProvincias = zona.getListaProvincias();
+            for (Provincia listaProvinciasProvincia : listaProvincias) {
+                listaProvinciasProvincia.setUnaZona(null);
+                listaProvinciasProvincia = em.merge(listaProvinciasProvincia);
             }
             em.remove(zona);
             em.getTransaction().commit();
@@ -109,4 +144,51 @@ public class ZonaJpaController implements Serializable{
             }
         }
     }
+
+    public List<Zona> findZonaEntities() {
+        return findZonaEntities(true, -1, -1);
+    }
+
+    public List<Zona> findZonaEntities(int maxResults, int firstResult) {
+        return findZonaEntities(false, maxResults, firstResult);
+    }
+
+    private List<Zona> findZonaEntities(boolean all, int maxResults, int firstResult) {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            cq.select(cq.from(Zona.class));
+            Query q = em.createQuery(cq);
+            if (!all) {
+                q.setMaxResults(maxResults);
+                q.setFirstResult(firstResult);
+            }
+            return q.getResultList();
+        } finally {
+            em.close();
+        }
+    }
+
+    public Zona findZona(int id) {
+        EntityManager em = getEntityManager();
+        try {
+            return em.find(Zona.class, id);
+        } finally {
+            em.close();
+        }
+    }
+
+    public int getZonaCount() {
+        EntityManager em = getEntityManager();
+        try {
+            CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
+            Root<Zona> rt = cq.from(Zona.class);
+            cq.select(em.getCriteriaBuilder().count(rt));
+            Query q = em.createQuery(cq);
+            return ((Long) q.getSingleResult()).intValue();
+        } finally {
+            em.close();
+        }
+    }
+    
 }
